Separación de resposabilidades: express.Router
