+++ routes/index.js
@@ -1,0 +1,11 @@
+const productsRouter = require('./products.router').router;
+const categoriesRouter = require('./categories.router');
+const usersRouter = require('./users.router');
+
+function routerApi(app) {
+  app.use('/products', productsRouter);
+  app.use('/categories', categoriesRouter);
+  app.use('/users', usersRouter);
+}
+
+module.exports = routerApi;
